move server components to App and add TODO placeholders
