+++ frontend/fingertips-frontend/app/Views/index.tsx
